signal: proper dnative split
Move dnative parts away from mainstream signal.c.
