+++ src/arch/linux/async/Makefile
@@ -6,8 +6,6 @@ ifeq ($(HAVE_LIBBFD),1)
 CFILES += backtrace-symbols.c
 endif
 
-ALL_CPPFLAGS += -I$(REALTOPDIR)/src/base/lib/mcontext
-
 include $(REALTOPDIR)/src/Makefile.common
 
 all: lib
