+++ src/arch/linux/async/signal.c
@@ -46,162 +46,6 @@
 #include "cpu-emu.h"
 #include "sig.h"
 
-/* sigaltstack work-around */
-#ifdef HAVE_LINUX_SIGNAL_H
-#include <linux/signal.h>
-#endif
-#define SIGALTSTACK_WA_DEFAULT DNATIVE
-#if SIGALTSTACK_WA_DEFAULT
-  #ifdef DISABLE_SYSTEM_WA
-    #ifdef SS_AUTODISARM
-      #define SIGALTSTACK_WA 0
-    #else
-      #ifdef WARN_UNDISABLED_WA
-        #warning Not disabling SIGALTSTACK_WA, update your kernel
-      #endif
-      #define SIGALTSTACK_WA 1
-    #endif
-  #else
-    /* work-around sigaltstack badness - disable when kernel is fixed */
-    #define SIGALTSTACK_WA 1
-  #endif
-  #if defined(WARN_OUTDATED_WA) && defined(SS_AUTODISARM)
-    #warning SIGALTSTACK_WA is outdated
-  #endif
-#else
-  #define SIGALTSTACK_WA 0
-#endif
-#if SIGALTSTACK_WA
-#include "mcontext.h"
-#include "mapping.h"
-#endif
-/* SS_AUTODISARM is a dosemu-specific sigaltstack extension supported
- * by some kernels */
-#ifndef SS_AUTODISARM
-#define SS_AUTODISARM  (1U << 31)    /* disable sas during sighandling */
-#endif
-
-/* sigreturn work-around */
-#ifdef HAVE_ASM_UCONTEXT_H
-#include <asm/ucontext.h>
-#endif
-#ifdef __x86_64__
-  #define SIGRETURN_WA_DEFAULT DNATIVE
-#else
-  #define SIGRETURN_WA_DEFAULT 0
-#endif
-#if SIGRETURN_WA_DEFAULT
-  #ifdef DISABLE_SYSTEM_WA
-    #ifdef UC_SIGCONTEXT_SS
-      #define SIGRETURN_WA 0
-    #else
-      #ifdef WARN_UNDISABLED_WA
-        #warning Not disabling SIGRETURN_WA, update your kernel
-      #endif
-      #define SIGRETURN_WA 1
-    #endif
-  #else
-    /* work-around sigreturn badness - disable when kernel is fixed */
-    #define SIGRETURN_WA 1
-  #endif
-  #if defined(WARN_OUTDATED_WA) && defined(UC_SIGCONTEXT_SS)
-    #warning SIGRETURN_WA is outdated
-  #endif
-#else
-  #define SIGRETURN_WA 0
-#endif
-
-#if SIGRETURN_WA
-static unsigned int *iret_frame;
-
-asm("\n\
-	.globl DPMI_iret\n\
-	.type DPMI_iret,@function\n\
-DPMI_iret:\n\
-	iretl\n\
-");
-extern void DPMI_iret(void);
-
-static void iret_frame_alloc(void)
-{
-    unsigned int i, j;
-    void *addr;
-    /* search for page with bits 16-31 clear within first 47 bits
-       of address space */
-    for (i = 1; i < 0x8000; i++) {
-        for (j = 0; j < 0x10000; j += PAGE_SIZE) {
-            addr = (void *) (i * 0x100000000UL + j);
-            iret_frame =
-                mmap_mapping(MAPPING_SCRATCH | MAPPING_NOOVERLAP,
-                             addr, PAGE_SIZE, PROT_READ | PROT_WRITE);
-            if (iret_frame != MAP_FAILED)
-                goto out;
-        }
-    }
-out:
-    if (iret_frame != addr) {
-        error("Can't find DPMI iret page, leaving\n");
-        config.exitearly = 1;
-    }
-}
-
-static void iret_frame_setup(sigcontext_t * scp)
-{
-    /* set up a frame to get back to DPMI via iret. The kernel does not save
-       %ss, and the SYSCALL instruction in sigreturn() destroys it.
-
-       IRET pops off everything in 64-bit mode even if the privilege
-       does not change which is nice, but clobbers the high 48 bits
-       of rsp if the DPMI client uses a 16-bit stack which is not so
-       nice (see EMUfailure.txt). Setting %rsp to 0x100000000 so that
-       bits 16-31 are zero works around this problem, as DPMI code
-       can't see bits 32-63 anyway.
-     */
-
-    iret_frame[0] = _scp_eip;
-    iret_frame[1] = _scp_cs;
-    iret_frame[2] = _scp_eflags;
-    iret_frame[3] = _scp_esp;
-    iret_frame[4] = _scp_ss;
-    _scp_rsp = (unsigned long) iret_frame;
-}
-
-static void dpmi_iret_setup(sigcontext_t * scp)
-{
-    iret_frame_setup(scp);
-    _scp_eflags &= ~TF;
-    _scp_rip = (unsigned long) DPMI_iret;
-    _scp_cs = getsegment(cs);
-}
-
-static void dpmi_iret_unwind(sigcontext_t * scp)
-{
-    if (_scp_rip != (unsigned long) DPMI_iret)
-        return;
-    _scp_eip = iret_frame[0];
-    _scp_cs = iret_frame[1];
-    _scp_eflags = iret_frame[2];
-    _scp_esp = iret_frame[3];
-    _scp_ss = iret_frame[4];
-}
-#endif
-
-#ifdef __x86_64__
-#ifndef UC_SIGCONTEXT_SS
-/*
- * UC_SIGCONTEXT_SS will be set when delivering 64-bit or x32 signals on
- * kernels that save SS in the sigcontext.  Kernels that set UC_SIGCONTEXT_SS
- * allow signal handlers to set UC_STRICT_RESTORE_SS;
- * if UC_STRICT_RESTORE_SS is set, then sigreturn will restore SS.
- *
- * For compatibility with old programs, the kernel will *not* set
- * UC_STRICT_RESTORE_SS when delivering signals from 32bit code.
- */
-#define UC_SIGCONTEXT_SS       0x2
-#define UC_STRICT_RESTORE_SS   0x4
-#endif
-#endif
-
 /* Variables for keeping track of signals */
 #define MAX_SIG_QUEUE_SIZE 50
 #define MAX_SIG_DATA_SIZE 128
@@ -240,15 +84,6 @@ struct callback_s {
 static sigset_t q_mask;
 static sigset_t nonfatal_q_mask;
 static sigset_t fatal_q_mask;
-static void *cstack;
-#if SIGALTSTACK_WA
-static void *backup_stack;
-static int need_sas_wa;
-#endif
-#if SIGRETURN_WA
-static int need_sr_wa;
-#endif
-static int block_all_sigs;
 static int sig_inited;
 
 static int sh_tid;
@@ -266,13 +101,11 @@ struct eflags_fs_gs eflags_fs_gs;
 static void (*sighandlers[NSIG])(sigcontext_t *, siginfo_t *);
 static void (*qsighandlers[NSIG])(int sig, siginfo_t *si, void *uc);
 static void (*asighandlers[NSIG])(void *arg);
-static struct sigaction sacts[NSIG];
 
 static void SIGALRM_call(void *arg);
 static void SIGIO_call(void *arg);
 static void sigasync(int sig, siginfo_t *si, void *uc);
 static void sigasync_std(int sig, siginfo_t *si, void *uc);
-static void leavedos_sig(int sig);
 
 static void _newsetqsig(int sig, void (*fun)(int sig, siginfo_t *si, void *uc))
 {
@@ -295,7 +128,7 @@ static void init_one_sig(int num, void (*fun)(int sig, siginfo_t *si, void *uc))
 	struct sigaction sa;
 
 	sa.sa_flags = SA_RESTART | SA_ONSTACK | SA_SIGINFO;
-	if (block_all_sigs)
+	if (signative_block_all_sigs())
 	{
 		/* initially block all async signals. */
 		sa.sa_mask = q_mask;
@@ -359,7 +192,7 @@ static void newsetsig(int sig, void (*fun)(int sig, siginfo_t *si, void *uc))
 	if (kernel_version_code >= KERNEL_VERSION(2, 6, 14))
 #endif
 		sa.sa_flags |= SA_NODEFER;
-	if (block_all_sigs)
+	if (signative_block_all_sigs())
 	{
 		/* initially block all async signals. */
 		sa.sa_mask = q_mask;
@@ -370,36 +203,6 @@ static void newsetsig(int sig, void (*fun)(int sig, siginfo_t *si, void *uc))
 		sa.sa_mask = nonfatal_q_mask;
 	}
 	sa.sa_sigaction = fun;
-	sigaction(sig, &sa, NULL);
-}
-
-SIG_PROTO_PFX
-static void fixup_handler(int sig, siginfo_t *si, void *uc)
-{
-	struct sigaction *sa;
-	ucontext_t *uct = uc;
-	sigcontext_t *scp = &uct->uc_mcontext;
-	init_handler(scp, 1);
-	sa = &sacts[sig];
-	if (sa->sa_flags & SA_SIGINFO) {
-		sa->sa_sigaction(sig, si, uc);
-	} else {
-		typedef void (*hdlr_t)(int, siginfo_t *, ucontext_t *);
-		hdlr_t hdlr = (hdlr_t)sa->sa_handler;
-		hdlr(sig, si, uc);
-	}
-	deinit_handler(scp, &uct->uc_flags);
-}
-
-static void fixupsig(int sig)
-{
-	struct sigaction sa;
-	sigaction(sig, NULL, &sa);
-	if (sa.sa_handler == SIG_DFL || sa.sa_handler == SIG_IGN)
-		return;
-	sacts[sig] = sa;
-	sa.sa_flags |= SA_ONSTACK | SA_SIGINFO;
-	sa.sa_sigaction = fixup_handler;
 	sigaction(sig, &sa, NULL);
 }
 
@@ -423,7 +226,7 @@ static void __init_handler(sigcontext_t *scp, unsigned long uc_flags)
      to save those ourselves */
   _scp_ds = getsegment(ds);
   _scp_es = getsegment(es);
-  if (!(uc_flags & UC_SIGCONTEXT_SS))
+  if (!signative_skip_ss(uc_flags))
     _scp_ss = getsegment(ss);
   _scp_fs = getsegment(fs);
   _scp_gs = getsegment(gs);
@@ -468,45 +271,7 @@ static void __init_handler(sigcontext_t *scp, unsigned long uc_flags)
     return;
   }
 
-#ifdef DNATIVE
-#if SIGRETURN_WA
-  if (need_sr_wa && !DPMIValidSelector(_scp_cs))
-    dpmi_iret_unwind(scp);
-#endif
-
-#if 0
-  /* for async signals need to restore fs/gs even if dosemu code
-   * was interrupted, because it can be interrupted in a switching
-   * routine when fs or gs are already switched but cs is not */
-  if (!DPMIValidSelector(_scp_cs) && !async)
-    return;
-#else
-  /* as DIRECT_DPMI_SWITCH support is now removed, the above comment
-   * applies only to DPMI_iret, which is now unwound.
-   * We don't need to restore segregs for async signals any more. */
-  if (!DPMIValidSelector(_scp_cs))
-    return;
-#endif
-
-  /* restore %fs and %gs for compatibility with NPTL. */
-  if (getsegment(fs) != eflags_fs_gs.fs)
-    loadregister(fs, eflags_fs_gs.fs);
-  if (getsegment(gs) != eflags_fs_gs.gs)
-    loadregister(gs, eflags_fs_gs.gs);
-#ifdef __x86_64__
-  loadregister(ds, eflags_fs_gs.ds);
-  loadregister(es, eflags_fs_gs.es);
-  /* kernel has the following rule: non-zero selector means 32bit base
-   * in GDT. Zero selector means 64bit base, set via msr.
-   * So if we set selector to 0, need to use also prctl(ARCH_SET_xS).
-   * Also, if the bases are not used they are 0 so no need to restore,
-   * which saves a syscall */
-  if (!eflags_fs_gs.fs && eflags_fs_gs.fsbase)
-    dosemu_arch_prctl(ARCH_SET_FS, eflags_fs_gs.fsbase);
-  if (!eflags_fs_gs.gs && eflags_fs_gs.gsbase)
-    dosemu_arch_prctl(ARCH_SET_GS, eflags_fs_gs.gsbase);
-#endif
-#endif
+  signative_enter(scp);
 }
 
 SIG_PROTO_PFX
@@ -530,15 +295,10 @@ void init_handler(sigcontext_t *scp, unsigned long uc_flags)
    * Sync signals like SIGSEGV are never blocked.
    */
   __init_handler(scp, uc_flags);
-  if (!block_all_sigs)
+  if (!signative_block_all_sigs())
     return;
-#if SIGALTSTACK_WA
-  /* for SAS WA we unblock the fatal signals even later if we came
-   * from DPMI, as then we'll be switching stacks which is racy when
-   * async signals enabled. */
-  if (need_sas_wa && DPMIValidSelector(_scp_cs))
+  if (signative_skip_unblock(scp))
     return;
-#endif
   /* either came from dosemu/vm86 or having SS_AUTODISARM -
    * then we can unblock any signals we want. This is because
    * dosemu DOES NOT USE signal stack by itself. We switch to
@@ -555,49 +315,13 @@ void init_handler(sigcontext_t *scp, unsigned long uc_flags)
 SIG_PROTO_PFX
 void deinit_handler(sigcontext_t *scp, unsigned long *uc_flags)
 {
-#ifdef DNATIVE
 #ifdef X86_EMULATOR
   /* in fullsim mode nothing to do */
   if (CONFIG_CPUSIM && EMU_DPMI())
     return;
 #endif
 
-  if (!DPMIValidSelector(_scp_cs))
-    return;
-
-#ifdef __x86_64__
-  if (*uc_flags & UC_SIGCONTEXT_SS) {
-    /*
-     * On Linux 4.4 (possibly) and up, the kernel can fully restore
-     * SS and ESP, so we don't need any special tricks.  To avoid confusion,
-     * force strict restore.  (Some 4.1 versions support this as well but
-     * without the uc_flags bits.  It's not trying to detect those kernels.)
-     */
-    *uc_flags |= UC_STRICT_RESTORE_SS;
-  } else {
-#if SIGRETURN_WA
-    if (!need_sr_wa) {
-      need_sr_wa = 1;
-      warn("Enabling sigreturn() work-around\n");
-    }
-    dpmi_iret_setup(scp);
-#else
-    error("Your kernel does not support UC_STRICT_RESTORE_SS and the "
-	  "work-around in dosemu is not enabled.\n");
-    dpmi_return(scp, DPMI_RET_EXIT);
-    leavedos_sig(-1);
-#endif
-  }
-
-  if (_scp_fs != getsegment(fs))
-    loadregister(fs, _scp_fs);
-  if (_scp_gs != getsegment(gs))
-    loadregister(gs, _scp_gs);
-
-  loadregister(ds, _scp_ds);
-  loadregister(es, _scp_es);
-#endif
-#endif
+  signative_leave(scp, uc_flags);
 }
 
 static void leavedos_call(void *arg)
@@ -713,7 +437,7 @@ void leavedos_from_sig(int sig)
   _leavedos_main(0, sig);
 }
 
-static void leavedos_sig(int sig)
+void leavedos_sig(int sig)
 {
   /* do not log anything from a sighandler or it may hang */
   SIGNAL_save(leavedos_call, &sig, sizeof(sig), __func__);
@@ -868,101 +592,6 @@ static void signal_thr(void *arg)
   sig_c.signal_handler(sig_c.arg);
 }
 
-static void sigstack_init(void)
-{
-#ifndef MAP_STACK
-#define MAP_STACK 0
-#endif
-
-  /* sigaltstack_wa is optional. See if we need it. */
-  /* .ss_flags is signed int and SS_AUTODISARM is a sign bit :( */
-  stack_t dummy2;
-  stack_t dummy = { .ss_flags = (int)(SS_DISABLE | SS_AUTODISARM) };
-  int err = sigaltstack(&dummy, &dummy2);
-  int errno_save = errno;
-
-  /* needs to drop SS_AUTODISARM or asan will fail. See
-   * https://github.com/dosemu2/dosemu2/issues/1576 */
-  sigaltstack(&dummy2, NULL);
-#if SIGALTSTACK_WA
-  if ((err && errno_save == EINVAL)
-#ifdef __i386__
-#ifdef __linux__
-      /* kernels before 4.11 had the needed functionality only for 64bits */
-      || kernel_version_code < KERNEL_VERSION(4, 11, 0)
-#endif
-#endif
-     )
-  {
-    need_sas_wa = 1;
-    warn("Enabling sigaltstack() work-around\n");
-    /* for SAS WA block all signals. If we don't, there is a
-     * race that the signal can come after we switched to backup stack
-     * but before we disabled sigaltstack. We unblock the fatal signals
-     * later, only right before switching back to dosemu. */
-    block_all_sigs = 1;
-  } else if (err) {
-    goto unk_err;
-  }
-
-  if (need_sas_wa) {
-    cstack = alloc_mapping(MAPPING_SHARED, SIGSTACK_SIZE);
-    if (cstack == MAP_FAILED) {
-      error("Unable to allocate stack\n");
-      config.exitearly = 1;
-      return;
-    }
-    backup_stack = alias_mapping_ux(MAPPING_OTHER, SIGSTACK_SIZE,
-	PROT_READ | PROT_WRITE, cstack);
-    if (backup_stack == MAP_FAILED) {
-      error("Unable to allocate stack\n");
-      config.exitearly = 1;
-      return;
-    }
-  } else {
-    cstack = mmap(NULL, SIGSTACK_SIZE, PROT_READ | PROT_WRITE,
-	MAP_PRIVATE | MAP_ANONYMOUS | MAP_STACK, -1, 0);
-    if (cstack == MAP_FAILED) {
-      error("Unable to allocate stack\n");
-      config.exitearly = 1;
-      return;
-    }
-  }
-#else
-  if ((err && errno_save == EINVAL)
-#ifdef __i386__
-#ifdef __linux__
-      || kernel_version_code < KERNEL_VERSION(4, 11, 0)
-#endif
-#endif
-     )
-  {
-    error("Your kernel does not support SS_AUTODISARM and the "
-	  "work-around in dosemu is not enabled.\n");
-    config.exitearly = 1;
-    return;
-  } else if (err) {
-    goto unk_err;
-  }
-  cstack = mmap(NULL, SIGSTACK_SIZE, PROT_READ | PROT_WRITE,
-	MAP_PRIVATE | MAP_ANONYMOUS | MAP_STACK, -1, 0);
-  if (cstack == MAP_FAILED) {
-    error("Unable to allocate stack\n");
-    config.exitearly = 1;
-    return;
-  }
-#endif
-
-  return;
-
-unk_err:
-  if (err) {
-    error("sigaltstack() returned %i, %s\n", errno_save,
-        strerror(errno_save));
-    config.exitearly = 1;
-  }
-}
-
 /* DANG_BEGIN_FUNCTION signal_pre_init
  *
  * description:
@@ -1015,9 +644,9 @@ signal_pre_init(void)
   setup_nf_sig(SIG_ACQUIRE);
   setup_nf_sig(SIG_RELEASE);
   setup_nf_sig(SIGWINCH);
+#ifdef DNATIVE
   fixupsig(SIGPROF);
   /* call that after all non-fatal sigs set up */
-#ifdef DNATIVE
   newsetsig(SIGILL, dosemu_fault);
   newsetsig(SIGFPE, dosemu_fault);
   newsetsig(SIGTRAP, dosemu_fault);
@@ -1034,34 +663,18 @@ signal_pre_init(void)
   prctl(PR_SET_PDEATHSIG, SIGQUIT);
 #endif
 
-#if SIGRETURN_WA
-  if (config.cpu_vm_dpmi == CPUVM_NATIVE)
-    iret_frame_alloc();
-#endif
-
   dosemu_pthread_self = pthread_self();
   rng_init(&cbks, MAX_CBKS, sizeof(struct callback_s));
+
+  signative_pre_init();
 }
 
 void
 signal_init(void)
 {
   /* signal_init is called after dpmi_setup so this check is safe */
-  if (config.cpu_vm_dpmi == CPUVM_NATIVE) {
-    sigstack_init();
-#if SIGRETURN_WA
-    /* 4.6+ are able to correctly restore SS */
-#ifdef __linux__
-    if (kernel_version_code < KERNEL_VERSION(4, 6, 0)) {
-      need_sr_wa = 1;
-      warn("Enabling sigreturn() work-around for old kernel\n");
-      /* block all sigs for SR WA. If we don't, the signal can come before
-       * SS is saved, but we can't restore SS on signal exit. */
-      block_all_sigs = 1;
-    }
-#endif
-#endif
-  }
+  if (config.cpu_vm_dpmi == CPUVM_NATIVE)
+    signative_init();
 
   sh_tid = coopth_create("signal handling", signal_thr);
   /* normally we don't need ctx handlers because the thread is detached.
@@ -1378,93 +991,6 @@ static void async_call(void *arg)
   process_callbacks();
 }
 
-#if SIGALTSTACK_WA
-static void signal_sas_wa(void)
-{
-  int err;
-  stack_t ss = {};
-  m_ucontext_t hack;
-  unsigned char *sp;
-  unsigned char *top = cstack + SIGSTACK_SIZE;
-  unsigned char *btop = backup_stack + SIGSTACK_SIZE;
-  ptrdiff_t delta;
-
-  if (getmcontext(&hack) == 0) {
-    sp = alloca(sizeof(void *));
-    delta = top - sp;
-    /* switch stack to its mirror (same phys addr) to cheat sigaltstack() */
-    asm volatile(
-#ifdef __x86_64__
-    "mov %0, %%rsp\n"
-#else
-    "mov %0, %%esp\n"
-#endif
-     :: "r"(btop - delta));
-  } else {
-    sigprocmask(SIG_UNBLOCK, &fatal_q_mask, NULL);
-    return;
-  }
-
-  ss.ss_flags = SS_DISABLE;
-  /* sas will re-enable itself when returning from sighandler */
-  err = sigaltstack(&ss, NULL);
-  if (err)
-    perror("sigaltstack");
-
-  setmcontext(&hack);
-}
-#endif
-
-#ifdef DNATIVE
-static int saved_fc;
-
-void signal_switch_to_dosemu(void)
-{
-  saved_fc = fault_cnt;
-  fault_cnt = 0;
-}
-
-void signal_switch_to_dpmi(void)
-{
-  fault_cnt = saved_fc;
-}
-
-void signal_return_to_dosemu(void)
-{
-#if SIGALTSTACK_WA
-  if (need_sas_wa)
-    signal_sas_wa();
-#endif
-}
-
-void signal_return_to_dpmi(void)
-{
-}
-
-void signal_set_altstack(int on)
-{
-  stack_t stk = { 0 };
-  int err;
-
-  if (!on) {
-    stk.ss_flags = SS_DISABLE;
-    err = sigaltstack(&stk, NULL);
-  } else {
-    stk.ss_sp = cstack;
-    stk.ss_size = SIGSTACK_SIZE;
-#if SIGALTSTACK_WA
-    stk.ss_flags = SS_ONSTACK | (need_sas_wa ? 0 : SS_AUTODISARM);
-#else
-    stk.ss_flags = SS_ONSTACK | SS_AUTODISARM;
-#endif
-    err = sigaltstack(&stk, NULL);
-  }
-  if (err) {
-    error("sigaltstack(0x%x) returned %i, %s\n",
-        stk.ss_flags, err, strerror(errno));
-    leavedos(err);
-  }
-}
 
 void signal_unblock_async_sigs(void)
 {
@@ -1479,7 +1005,12 @@ void signal_restore_async_sigs(void)
    * interrupted after changing %fs, we are in troubles */
   sigprocmask(SIG_BLOCK, &nonfatal_q_mask, NULL);
 }
-#endif
+
+void signal_unblock_fatal_sigs(void)
+{
+  /* unblock only nonfatal, fatals should already be unblocked */
+  sigprocmask(SIG_UNBLOCK, &fatal_q_mask, NULL);
+}
 
 /* similar to above, but to call from non-signal context */
 void signal_block_async_nosig(sigset_t *old_mask)
