+++ src/dosext/dpmi/dnative/Makefile
@@ -3,7 +3,8 @@ include $(top_builddir)/Makefile.conf
 
 CFILES = dnative.c sigsegv.c
 
-ALL_CPPFLAGS += -I$(REALTOPDIR)/src/dosext/dpmi
+ALL_CPPFLAGS += -I$(REALTOPDIR)/src/dosext/dpmi \
+  -I$(REALTOPDIR)/src/base/lib/mcontext
 
 all: lib
